Changes to remove version id from getPastAlerts
So on getPastAlerts: because the past alerts object doesn't contain the version ID, we will only see the query result as (repeating) alert IDS. So even though this morning we decided to return the alert ID and version ID, we actually need to only return the distinct alert IDs to avoid creating duplicate objects. This script should return all alert IDs that fall within the from_time and to_time period.

And then for getPastAlertsVersions: Because we are only returning the alert_id from getPastAlerts, this script needs to return all the relevant versions for an alert_id. So we are thinking that it should take in four parameters: alert_id, include_all_versions, from_time and to_time. We need the latter two because this script now has to include the logic to return alert versions that fall within the time frame. The if statements concerning include_all_versions currently in getPastAlerts would also then be moved over to getPastAlertsVersions. This script would still return the same fields.